Extract EmojiPicker
